+++ CHANGES.rst
@@ -2,6 +2,46 @@
 Changelog
 ==========
 
+0.2.0 (2025-01-01) 🎉🎉🎉 HAPPY NEW YEAR 🎉🎉🎉
+==================
+
+Features
+--------
+
+- Add AsyncSQLite.py
+- more flexible and advanced functionality
+- add request WHERE
+- add request LIMIT
+- add request ORDER BY
+- add request UPDATE
+- add request DELETE
+- add async EXECUTE
+- add async EXECUTEMANY
+- add async FETCH ONE
+
+Misc
+--------
+
+- Fixed Bug
+- more secure queries from SQL-injections
+
+
+
+0.1.3 (2024-12-21)
+==================
+
+Debug
+--------
+
+- Insert SQL-request
+
+Misc
+-----
+
+- Fixed design of the project
+- Added __meta__.py
+- Added directory .client
+
 
 0.1.2 (2024-12-13)
 ==================